+++ sktime/detection/adapters/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3 -u
 # copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
-"""Implements adapters for time series annotation."""
+"""Implements adapters for time series detection."""
 
-__all__ = ["PyODAnnotator"]
+__all__ = ["PyODAnnotator", "PyODDetector"]
 
-from sktime.annotation.adapters._pyod import PyODAnnotator
+from sktime.detection.adapters._pyod import PyODAnnotator, PyODDetector
